+++ src/features/auth/LoginButton/LoginButton.tsx
@@ -1,6 +1,6 @@
 function LoginButton() {
   return (
-    <button className='btn btn-sm btn-accent md:btn-md rounded-md tracking-wider'>
+    <button className='btn btn-sm btn-accent md:btn-md tracking-wider'>
       Login
     </button>
   );