+++ src/features/auth/SignUpButton/SignUpButton.tsx
@@ -1,6 +1,6 @@
 function SignUpButton() {
   return (
-    <button className='btn btn-sm btn-accent btn-outline md:btn-md rounded-md tracking-wider'>
+    <button className='btn btn-sm btn-accent btn-outline md:btn-md tracking-wider'>
       Sign up
     </button>
   );