+++ src/features/offers/SavedOffers.tsx
@@ -1,0 +1,5 @@
+function SavedOffers() {
+  return <section></section>;
+}
+
+export default SavedOffers;