+++ src/features/offers/index.ts
@@ -1,0 +1,3 @@
+import SavedOffers from './SavedOffers';
+
+export default SavedOffers;